Move ExtraPeers to standalone module
This is a Cardano specific type and this is a step to remove any
lingering cardano-node details from the network code and move them to
the cardano-node sublibrary.
