+++ ouroboros-network/src/Ouroboros/Cardano/Network/PublicRootPeers.hs
@@ -1,102 +1,132 @@
-{-# LANGUAGE NamedFieldPuns #-}
-
+{-# LANGUAGE FlexibleInstances    #-}
+{-# LANGUAGE NamedFieldPuns       #-}
+{-# LANGUAGE TypeSynonymInstances #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 module Ouroboros.Cardano.Network.PublicRootPeers where
 
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
 import Data.Set (Set)
 import Data.Set qualified as Set
+import Ouroboros.Cardano.Network.ExtraRootPeers (ExtraPeers (..))
+import Ouroboros.Cardano.Network.ExtraRootPeers qualified as ExtraPeers
 import Ouroboros.Network.PeerSelection.PeerAdvertise (PeerAdvertise (..))
-import Ouroboros.Network.PeerSelection.Types
+import Ouroboros.Network.PeerSelection.PublicRootPeers
 
-data ExtraPeers peeraddr =
-  ExtraPeers
-    { getPublicConfigPeers :: !(Map peeraddr PeerAdvertise)
-    , getBootstrapPeers    :: !(Set peeraddr)
-    }
-  deriving (Eq, Show)
+type CardanoPublicRootPeers peeraddr =
+  PublicRootPeers (ExtraPeers peeraddr) peeraddr
 
-instance Ord peeraddr => Semigroup (ExtraPeers peeraddr) where
-  (ExtraPeers a b) <> (ExtraPeers a' b') =
-    let -- Combine the sets, prioritizing bootstrapPeers
-        combinedSet = b `Set.union` b'
-        -- Combine the maps and remove any peers that are now in the set
-        combinedMap = (a `Map.union` a') `Map.withoutKeys` combinedSet
-    in ExtraPeers combinedMap combinedSet
+instance ( Ord peeraddr
+         ) => Semigroup (CardanoPublicRootPeers peeraddr) where
+  (<>) = merge
 
-instance Ord peeraddr => Monoid (ExtraPeers peeraddr) where
-  mempty = empty
+instance ( Ord peeraddr
+         ) => Monoid (CardanoPublicRootPeers peeraddr) where
+  mempty = empty ExtraPeers.empty
 
--- Cardano Public Root Peers Actions
-cardanoPublicRootPeersAPI :: Ord peeraddr => PublicExtraPeersAPI (ExtraPeers peeraddr) peeraddr
-cardanoPublicRootPeersAPI =
-  PublicExtraPeersAPI {
-    nullExtraPeers         = nullAll
-  , invariantExtraPeers    = invariant
-  , memberExtraPeers       = member
-  , extraPeersToSet        = toSet
-  , sizeExtraPeers         = size
-  , differenceExtraPeers   = difference
-  , intersectionExtraPeers = intersection
-  , toAdvertise            = toAdvertisePeersMap
-  }
+merge :: Ord peeraddr
+      => CardanoPublicRootPeers peeraddr
+      -> CardanoPublicRootPeers peeraddr
+      -> CardanoPublicRootPeers peeraddr
+merge a b = mergeG ExtraPeers.toSet a b
 
--- Map and Set are disjoint
+getPublicConfigPeers :: CardanoPublicRootPeers peeraddr -> Map peeraddr PeerAdvertise
+getPublicConfigPeers PublicRootPeers { getExtraPeers = (ExtraPeers pp _) } = pp
+
+getBootstrapPeers :: CardanoPublicRootPeers peeraddr -> Set peeraddr
+getBootstrapPeers PublicRootPeers { getExtraPeers = (ExtraPeers _ bsp) } = bsp
+
+toPublicConfigPeerSet :: CardanoPublicRootPeers peeraddr -> Set peeraddr
+toPublicConfigPeerSet PublicRootPeers {
+                        getExtraPeers = ExtraPeers pp _
+                      } = Map.keysSet pp
+
+-- | Preserves PublicRootPeers invariant. If the two sets are not disjoint,
+-- removes the common ones from the bootstrap peers set since its the most
+-- sensitive set.
 --
-invariant :: Ord peeraddr => ExtraPeers peeraddr -> Bool
-invariant (ExtraPeers a b) = all (`Map.notMember` a) b
-
 fromMapAndSet :: Ord peeraddr
-              => Map peeraddr PeerAdvertise
-              -> Set peeraddr
-              -> ExtraPeers peeraddr
+              => Map peeraddr PeerAdvertise -- ^ public configured root peers
+              -> Set peeraddr -- ^ bootstrap peers
+              -> Set peeraddr -- ^ ledger peers
+              -> Set peeraddr -- ^ big ledger peers
+              -> CardanoPublicRootPeers peeraddr
 fromMapAndSet pp bsp =
-  let newPP = pp `Map.withoutKeys` bsp
-   in ExtraPeers newPP bsp
+  fromDisjointSets ExtraPeers.toSet (ExtraPeers.fromMapAndSet pp bsp)
 
-empty :: ExtraPeers peeraddr
-empty = ExtraPeers Map.empty Set.empty
+fromPublicRootPeers :: Map peeraddr PeerAdvertise
+                    -> CardanoPublicRootPeers peeraddr
+fromPublicRootPeers pp =
+  empty (ExtraPeers.empty { publicConfigPeers = pp })
 
-nullPublicConfig :: ExtraPeers peeraddr -> Bool
-nullPublicConfig ExtraPeers { getPublicConfigPeers } =
-  Map.null getPublicConfigPeers
+fromBootstrapPeers :: Set peeraddr
+                   -> CardanoPublicRootPeers peeraddr
+fromBootstrapPeers bsp =
+  empty (ExtraPeers.empty { bootstrapPeers = bsp })
 
-nullBootstrap :: ExtraPeers peeraddr -> Bool
-nullBootstrap ExtraPeers { getBootstrapPeers } =
-  Set.null getBootstrapPeers
+insertPublicConfigPeer :: Ord peeraddr
+                       => peeraddr
+                       -> PeerAdvertise
+                       -> CardanoPublicRootPeers peeraddr
+                       -> CardanoPublicRootPeers peeraddr
+insertPublicConfigPeer p pa prp@PublicRootPeers {
+                              getExtraPeers = ExtraPeers {
+                                bootstrapPeers = getBootstrapPeers'
+                              }
+                            } =
+  let prp'@PublicRootPeers {
+        getExtraPeers = ExtraPeers {
+          publicConfigPeers = pp
+        }
+      } = difference ExtraPeers.difference prp (Set.singleton p)
+  in if Set.member p getBootstrapPeers'
+     then prp
+     else prp' { getExtraPeers = (getExtraPeers prp') {
+                   publicConfigPeers = Map.insert p pa pp
+                 }
+               }
 
-nullAll :: ExtraPeers peeraddr -> Bool
-nullAll cprp = nullPublicConfig cprp && nullBootstrap cprp
+insertBootstrapPeer :: Ord peeraddr
+                    => peeraddr
+                    -> CardanoPublicRootPeers peeraddr
+                    -> CardanoPublicRootPeers peeraddr
+insertBootstrapPeer p prp =
+  let prp'@PublicRootPeers {
+        getExtraPeers = ExtraPeers {
+          bootstrapPeers = bsp
+        }
+      } = difference ExtraPeers.difference prp (Set.singleton p)
+  in prp' { getExtraPeers = (getExtraPeers prp') {
+              bootstrapPeers = Set.insert p bsp
+            }
+          }
 
-member :: Ord peeraddr => peeraddr -> ExtraPeers peeraddr -> Bool
-member addr (ExtraPeers a b) =
-  Map.member addr a || Set.member addr b
+insertLedgerPeer :: Ord peeraddr
+                 => peeraddr
+                 -> CardanoPublicRootPeers peeraddr
+                 -> CardanoPublicRootPeers peeraddr
+insertLedgerPeer p prp@PublicRootPeers {
+                     getExtraPeers = ExtraPeers {
+                       bootstrapPeers = getBootstrapPeers'
+                     }
+                   } =
+  let prp'@PublicRootPeers { getLedgerPeers } =
+        difference ExtraPeers.difference prp (Set.singleton p)
+   in if Set.member p getBootstrapPeers'
+         then prp
+         else prp' {getLedgerPeers = Set.insert p getLedgerPeers }
 
-toSet :: Ord peeraddr => ExtraPeers peeraddr -> Set peeraddr
-toSet (ExtraPeers a b) = Map.keysSet a <> b
-
-size :: ExtraPeers peeraddr -> Int
-size (ExtraPeers a b) = Map.size a + Set.size b
-
-difference :: Ord peeraddr
-           => ExtraPeers peeraddr
-           -> Set peeraddr
-           -> ExtraPeers peeraddr
-difference (ExtraPeers a b) addrs =
-  ExtraPeers (Map.withoutKeys a addrs)
-                         (Set.difference b addrs)
-
-
-intersection :: Ord peeraddr
-             => ExtraPeers peeraddr
-             -> Set peeraddr
-             -> ExtraPeers peeraddr
-intersection (ExtraPeers a b) addrs =
-  ExtraPeers (Map.restrictKeys a addrs)
-                         (Set.intersection b addrs)
-
-toAdvertisePeersMap :: Ord peeraddr
-                    => ExtraPeers peeraddr
-                    -> Map peeraddr PeerAdvertise
-toAdvertisePeersMap (ExtraPeers a b) =
-  a <> Set.foldl (\m p -> Map.insert p DoNotAdvertisePeer m) Map.empty b
+insertBigLedgerPeer :: Ord peeraddr
+                    => peeraddr
+                    -> CardanoPublicRootPeers peeraddr
+                    -> CardanoPublicRootPeers peeraddr
+insertBigLedgerPeer p prp@PublicRootPeers{
+                        getExtraPeers = ExtraPeers {
+                          bootstrapPeers = getBootstrapPeers'
+                        }
+                      } =
+  let prp'@PublicRootPeers { getBigLedgerPeers } =
+        difference ExtraPeers.difference prp (Set.singleton p)
+   in if Set.member p getBootstrapPeers'
+         then prp
+         else prp' { getBigLedgerPeers = Set.insert p getBigLedgerPeers }
